--- GestionDeEmpleados/src/main/java/org/example/exception/EmailDuplicadoException.java
@@ -1,5 +1,0 @@
-package org.example.exception;
-
-public class EmailDuplicadoException extends RuntimeException {
-    public EmailDuplicadoException(String message) { super(message); }
-}
