--- GestionDeEmpleados/src/main/java/org/example/exception/RecursoNoEncontradoException.java
@@ -1,5 +1,0 @@
-package org.example.exception;
-
-public class RecursoNoEncontradoException extends RuntimeException {
-    public RecursoNoEncontradoException(String message) { super(message); }
-}
