+++ GestionDeEmpleados/src/main/java/org/example/gestiondeempleados/GestionDeEmpleadosApplication.java
@@ -1,0 +1,13 @@
+package org.example.gestiondeempleados;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GestionDeEmpleadosApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GestionDeEmpleadosApplication.class, args);
+	}
+
+}
