+++ GestionDeEmpleados/src/main/java/org/example/gestiondeempleados/controller/EmpleadoController.java
@@ -1,0 +1,59 @@
+package org.example.gestiondeempleados.controller;
+
+import jakarta.validation.Valid;
+import org.example.gestiondeempleados.model.Empleado;
+import org.example.gestiondeempleados.service.EmpleadoService;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import java.math.BigDecimal;
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/empleados")
+public class EmpleadoController {
+
+    private final EmpleadoService empleadoService;
+
+    public EmpleadoController(EmpleadoService empleadoService) {
+        this.empleadoService = empleadoService;
+    }
+
+    @GetMapping
+    public List<Empleado> obtenerTodos() {
+        return empleadoService.obtenerTodos();
+    }
+
+    @GetMapping("/{id}")
+    public Empleado obtenerPorId(@PathVariable Long id) {
+        return empleadoService.buscarPorId(id);
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public Empleado crear(@Valid @RequestBody Empleado empleado) {
+        return empleadoService.guardar(empleado);
+    }
+
+    @PutMapping("/{id}")
+    public Empleado actualizar(@PathVariable Long id, @Valid @RequestBody Empleado empleado) {
+        return empleadoService.actualizar(id, empleado);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void eliminar(@PathVariable Long id) {
+        empleadoService.eliminar(id);
+    }
+
+    @GetMapping("/departamento/{nombre}")
+    public List<Empleado> obtenerPorDepartamento(@PathVariable String nombre) {
+        return empleadoService.buscarPorDepartamento(nombre);
+    }
+
+    @GetMapping("/salario")
+    public List<Empleado> obtenerPorRangoSalario(@RequestParam BigDecimal min,
+                                                 @RequestParam BigDecimal max) {
+        return empleadoService.buscarPorRangoSalario(min, max);
+    }
+}
