+++ GestionDeEmpleados/src/main/java/org/example/gestiondeempleados/exception/EmailDuplicadoException.java
@@ -1,0 +1,5 @@
+package org.example.gestiondeempleados.exception;
+
+public class EmailDuplicadoException extends RuntimeException {
+    public EmailDuplicadoException(String message) { super(message); }
+}
