+++ GestionDeEmpleados/src/main/java/org/example/gestiondeempleados/exception/RecursoNoEncontradoException.java
@@ -1,0 +1,5 @@
+package org.example.gestiondeempleados.exception;
+
+public class RecursoNoEncontradoException extends RuntimeException {
+    public RecursoNoEncontradoException(String message) { super(message); }
+}
