+++ GestionDeEmpleados/src/main/java/org/example/gestiondeempleados/service/EmpleadoService.java
@@ -1,0 +1,19 @@
+package org.example.gestiondeempleados.service;
+
+
+
+import org.example.gestiondeempleados.model.Empleado;
+
+import java.math.BigDecimal;
+import java.util.List;
+
+public interface EmpleadoService {
+    Empleado guardar(Empleado empleado);
+    Empleado buscarPorId(Long id);
+    List<Empleado> buscarPorDepartamento(String nombreDepartamento);
+    List<Empleado> buscarPorRangoSalario(BigDecimal salarioMin, BigDecimal salarioMax);
+    BigDecimal obtenerSalarioPromedioPorDepartamento(Long departamentoId);
+    List<Empleado> obtenerTodos();
+    Empleado actualizar(Long id, Empleado empleado);
+    void eliminar(Long id);
+}
