--- GestionDeEmpleados/src/main/java/org/example/model/Departamento.java
@@ -1,32 +1,0 @@
-package org.example.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.*;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "departamentos",
-        uniqueConstraints = @UniqueConstraint(name = "uk_departamento_nombre", columnNames = "nombre"))
-@Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class Departamento {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotBlank
-    @Size(max = 100)
-    @Column(nullable = false, length = 100)
-    private String nombre;
-
-    @Size(max = 500)
-    @Column(length = 500)
-    private String descripcion;
-
-    @OneToMany(mappedBy = "departamento", cascade = CascadeType.ALL, orphanRemoval = false)
-    @Builder.Default
-    private List<Empleado> empleados = new ArrayList<>();
-}
