--- GestionDeEmpleados/src/main/java/org/example/repository/DepartamentoRepository.java
@@ -1,11 +1,0 @@
-package org.example.repository;
-
-import org.example.model.Departamento;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface DepartamentoRepository extends JpaRepository<Departamento, Long> {
-    Optional<Departamento> findByNombre(String nombre);
-    boolean existsByNombre(String nombre);
-}
