--- GestionDeEmpleados/src/main/java/org/example/service/DepartamentoService.java
@@ -1,15 +1,0 @@
-package org.example.service;
-
-import org.example.model.Departamento;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-public interface DepartamentoService {
-    Departamento guardar(Departamento dpto);
-    Departamento buscarPorId(Long id);
-    List<Departamento> obtenerTodos();
-    Departamento actualizar(Long id, Departamento dpto);
-    void eliminar(Long id);
-    BigDecimal salarioPromedio(Long departamentoId);
-}
