+++ GestionDeEmpleados/src/test/java/org/example/gestiondeempleados/GestionDeEmpleadosApplicationTests.java
@@ -1,0 +1,13 @@
+package org.example.gestiondeempleados;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GestionDeEmpleadosApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
